MERA-52: (chore) Refactor namespaces and update package dependencies in MeraStore.Shared.Kernel.Persistence
