+++ src/MeraStore.Shared.Kernel.Persistence/GlobalUsings.cs
@@ -1,0 +1,4 @@
+// Global using directives
+
+global using System.Diagnostics.CodeAnalysis;
+global using Microsoft.EntityFrameworkCore;
