+++ src/MeraStore.Shared.Kernel.Persistence/Repositories/Repository.cs
@@ -1,8 +1,6 @@
 ﻿using System.Linq.Expressions;
 
 using MeraStore.Shared.Kernel.Persistence.Interfaces;
-
-using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Query;
 
 namespace MeraStore.Shared.Kernel.Persistence.Repositories
